Move OrganisationContext to /contexts and remove the useOrganisations.ts, which was only a forward function
