+++ src/components/auth/ProtectedRoute.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useAuth } from '../../auth';
 import { useConfig } from '../../contexts/configContext.ts';
 import LoginRedirect from '../../auth/LoginRedirect';
-import { useOrganisations } from '../../data/organisations/hooks/useOrganisations.ts';
+import { useOrganisationsContext } from '../../contexts/useOrganisationsContext.ts';
 import { Navigate } from 'react-router-dom';
 
 interface ProtectedRouteProps {
@@ -12,7 +12,7 @@ interface ProtectedRouteProps {
 export const ProtectedRoute: React.FC<ProtectedRouteProps> = ({ element }) => {
   const { isAuthenticated, isLoading } = useAuth();
   const { oidcConfig } = useConfig();
-  const { currentOrganisation, loading: organisationsLoading } = useOrganisations();
+  const { currentOrganisation, loading: organisationsLoading } = useOrganisationsContext();
 
   if (!oidcConfig) {
     return element;
