+++ src/data/vehicles/hooks/useVehicle.ts
@@ -3,7 +3,7 @@ import { useConfig } from '../../../contexts/configContext.ts';
 import { useAuth } from '../../../auth/authUtils.ts';
 import type { VehicleGQLShaped } from '../types/vehicleGqlShaped.ts';
 import { fetchVehicle } from '../api/fetchVehicles.ts';
-import { useOrganisations } from '../../organisations/hooks/useOrganisations.ts';
+import { useOrganisationsContext } from '../../../contexts/useOrganisationsContext.ts';
 
 /**
  * Fetch and parse a single VehicleGQLShaped from Sobek's GraphQL endpoint.
@@ -16,7 +16,7 @@ export function useVehicle(id: string | undefined) {
   const [error, setError] = useState<string | null>(null);
   const { applicationBaseUrl } = useConfig();
   const { getAccessToken } = useAuth();
-  const { currentOrganisation } = useOrganisations();
+  const { currentOrganisation } = useOrganisationsContext();
 
   const doFetch = useCallback(async () => {
     if (!id) return;
